fix: LoginViewModel에서 SocialLogin enum 분리
